+++ FinModelUtility/Fin/src/math/matrix/MatrixTransformUtil.cs
@@ -1,6 +1,6 @@
 ﻿using System.Numerics;
-
 using fin.model;
+
 
 namespace fin.math.matrix {
   public static class MatrixTransformUtil {
@@ -39,20 +39,20 @@ namespace fin.math.matrix {
 
       var matrix = new FinMatrix4x4();
 
-      matrix[0, 0] = 1.0 - 2.0 * qy * qy - 2.0 * qz * qz;
-      matrix[0, 1] = 2.0 * qx * qy - 2.0 * qz * qw;
-      matrix[0, 2] = 2.0 * qx * qz + 2.0 * qy * qw;
-      matrix[0, 3] = 0.0;
+      matrix[0, 0] = (float)(1.0 - 2.0 * qy * qy - 2.0 * qz * qz);
+      matrix[0, 1] = (float)(2.0 * qx * qy - 2.0 * qz * qw);
+      matrix[0, 2] = (float)(2.0 * qx * qz + 2.0 * qy * qw);
+      matrix[0, 3] = 0.0f;
 
-      matrix[1, 0] = 2.0 * qx * qy + 2.0 * qz * qw;
-      matrix[1, 1] = 1.0 - 2.0 * qx * qx - 2.0 * qz * qz;
-      matrix[1, 2] = 2.0 * qy * qz - 2.0 * qx * qw;
-      matrix[1, 3] = 0.0;
+      matrix[1, 0] = (float)(2.0 * qx * qy + 2.0 * qz * qw);
+      matrix[1, 1] = (float)(1.0 - 2.0 * qx * qx - 2.0 * qz * qz);
+      matrix[1, 2] = (float)(2.0 * qy * qz - 2.0 * qx * qw);
+      matrix[1, 3] = 0.0f;
 
-      matrix[2, 0] = 2.0 * qx * qz - 2.0 * qy * qw;
-      matrix[2, 1] = 2.0 * qy * qz + 2.0 * qx * qw;
-      matrix[2, 2] = 1.0 - 2.0 * qx * qx - 2.0 * qy * qy;
-      matrix[2, 3] = 0.0;
+      matrix[2, 0] = (float)(2.0 * qx * qz - 2.0 * qy * qw);
+      matrix[2, 1] = (float)(2.0 * qy * qz + 2.0 * qx * qw);
+      matrix[2, 2] = (float)(1.0 - 2.0 * qx * qx - 2.0 * qy * qy);
+      matrix[2, 3] = 0.0f;
 
       matrix[3, 0] = 0;
       matrix[3, 1] = 0;
@@ -66,10 +66,7 @@ namespace fin.math.matrix {
 
     public static IFinMatrix4x4 FromScale(IScale scale) {
       var matrix = new FinMatrix4x4 {
-          [0, 0] = scale.X,
-          [1, 1] = scale.Y,
-          [2, 2] = scale.Z,
-          [3, 3] = 1,
+          [0, 0] = scale.X, [1, 1] = scale.Y, [2, 2] = scale.Z, [3, 3] = 1,
       };
       matrix.UpdateState();
       return matrix;
